Fixing some update gcc.

+++ src/proof/pdr/pdrInv.c
@@ -670,7 +670,7 @@ Vec_Str_t * Pdr_InvPrintStr( Vec_Int_t * vInv, Vec_Int_t * vCounts )
     Vec_Int_t * vMap = Pdr_InvMap( vCounts );
     int nVars = Vec_IntSize(vCounts) - Vec_IntCountZero(vCounts);
     int i, k, * pCube, * pList = Vec_IntArray(vInv);
-    char * pBuffer = ABC_ALLOC( char, nVars );
+    char * pBuffer = ABC_ALLOC( char, (size_t)(unsigned)nVars );
     for ( i = 0; i < nVars; i++ )
         pBuffer[i] = '-';
     Pdr_ForEachCube( pList, pCube, i )
